server/requirements.txt
uvicorn
fastapi
requests
python-dotenv
pymongo